+++ studio-ui/ui/app/src/state/actions/pathNavigator.ts
@@ -60,6 +60,7 @@ export const pathNavigatorConditionallySetPathComplete = /*#__PURE__*/ createAct
 
 export const pathNavigatorConditionallySetPathFailed = /*#__PURE__*/ createAction<{
 	id: string;
+	path: string;
 	error: { status: number; message: string };
 }>('PATH_NAV_CONDITIONALLY_SET_PATH_FAILED');